+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 1.3.0 (not released)
 ~~~~~~~~~~~~~~~~~~~~
 
+* DateInput widget renders hardcoded "%Y-%m-%d" format
+
 * Adding ``supports_microseconds`` attribute to all relevant widget classes.
   Thanks to Stephen Burrows for the patch.